hotfix(DSLY-17, #50): Prevent duplicate likes for an artwork.
This commit fixes the issue where users could submit multiple likes for the same artwork. It adds a check in the createLike method to verify if the user has already liked the artwork before saving a new like. If the user has already liked the artwork, a BadRequestAlertException is thrown with an appropriate error message. This prevents duplicate likes and ensures that each user can only like an artwork once.