feat(DSLY-25): Add new feature to ArtworkService
This commit adds a new feature to the ArtworkService class. The save method has been implemented to handle saving artwork information. It converts the artworkDTO object to an entity, sets the status to "OK," saves it in the repository, and returns the converted DTO. This feature enhances the functionality of the ArtworkService by enabling the saving of artwork data.

.